requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0
openai>=1.0.0
groq>=0.4.0
requests>=2.28.0
aiohttp>=3.8.0
pyyaml>=6.0
pytest>=7.0
pytest-asyncio>=0.20
python-dotenv>=1.0.0
httpx>=0.24.0
duckduckgo-search>=6.0.0
